requirements.txt
Flask==2.3.3
pytest==7.4.2
pytest-cov
pytest-mock
pytest-playwright
requests

